% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metricas_score_based.R
\name{auc_pr}
\alias{auc_pr}
\title{Calcular AUC-PR}
\usage{
auc_pr(y_true, y_score)
}
\arguments{
\item{y_true}{Vector con valores reales codificados como 0 y 1.}

\item{y_score}{Vector con scores o probabilidades predichas.}
}
\value{
Lista con statistic y se.
}
\description{
Calcula el área bajo la curva precision-recall mediante integración trapezoidal.
}
\examples{
y_true <- c(1, 1, 0, 0)
y_score <- c(0.9, 0.7, 0.4, 0.2)
auc_pr(y_true, y_score)

}
