% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliares.R
\name{bootstrap_se_2x2}
\alias{bootstrap_se_2x2}
\title{Estimar error estandar mediante bootstrap}
\usage{
bootstrap_se_2x2(tab, stat_fun, B = 1000, seed = 123)
}
\arguments{
\item{tab}{Matriz o tabla 2x2.}

\item{stat_fun}{Funcion que recibe una tabla 2x2 y devuelve un estadistico numerico.}

\item{B}{Numero de remuestreos bootstrap.}

\item{seed}{Semilla para reproducibilidad.}
}
\value{
Error estandar bootstrap.
}
\description{
Estima el error estandar de una métrica calculada sobre una tabla 2x2
mediante remuestreo bootstrap.
}
