% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metricas_score_based.R
\name{brier_score}
\alias{brier_score}
\title{Calcular el Brier score}
\usage{
brier_score(y_true, y_score)
}
\arguments{
\item{y_true}{Vector con valores reales codificados como 0 y 1.}

\item{y_score}{Vector con scores o probabilidades predichas.}
}
\value{
Lista con statistic y se.
}
\description{
Calcula el Brier score, definido como el error cuadrático medio entre
las probabilidades predichas y los valores reales.
}
\examples{
y_true <- c(1, 1, 0, 0)
y_score <- c(0.9, 0.7, 0.4, 0.2)
brier_score(y_true, y_score)

}
