% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliares.R
\name{check_2x2}
\alias{check_2x2}
\title{Comprobar tabla 2x2}
\usage{
check_2x2(tab)
}
\arguments{
\item{tab}{Matriz o tabla 2x2.}
}
\value{
TRUE de forma invisible si la tabla es valida.
}
\description{
Comprueba que la entrada es una matriz o tabla de dimension 2x2,
sin valores perdidos y sin frecuencias negativas.
}
