% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metricas_score_based.R
\name{check_binary_vectors}
\alias{check_binary_vectors}
\title{Comprobar vectores binarios y scores}
\usage{
check_binary_vectors(y_true, y_score)
}
\arguments{
\item{y_true}{Vector con valores reales codificados como 0 y 1.}

\item{y_score}{Vector con scores o probabilidades predichas.}
}
\value{
TRUE de forma invisible si las comprobaciones son correctas.
}
\description{
Comprueba que los vectores de valores reales y scores tienen la misma longitud,
que no contienen valores perdidos y que la variable real es binaria.
}
