% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliares.R
\name{df_to_2x2}
\alias{df_to_2x2}
\title{Convertir datos binarios a tabla 2x2}
\usage{
df_to_2x2(real, pred, positive = 1)
}
\arguments{
\item{real}{Vector con valores reales.}

\item{pred}{Vector con valores predichos.}

\item{positive}{Valor que se considera clase positiva.}
}
\value{
Matriz 2x2 con la convencion TP, FN, FP, TN.
}
\description{
Construye una tabla 2x2 a partir de dos vectores: valores reales y predichos.
}
\examples{
real <- c(1, 1, 0, 0)
pred <- c(1, 0, 1, 0)
df_to_2x2(real, pred)

}
