% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliares.R
\name{get_cells}
\alias{get_cells}
\title{Extraer celdas de una tabla 2x2}
\usage{
get_cells(tab)
}
\arguments{
\item{tab}{Matriz o tabla 2x2.}
}
\value{
Lista con TP, FN, FP y TN.
}
\description{
Extrae TP, FN, FP y TN de una tabla 2x2 siguiendo la convencion del paquete.
}
\examples{
tab <- matrix(c(50, 10,
                5, 35),
              nrow = 2,
              byrow = TRUE)
get_cells(tab)

}
