% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliares.R
\name{get_N}
\alias{get_N}
\title{Calcular tamano muestral total}
\usage{
get_N(tab)
}
\arguments{
\item{tab}{Matriz o tabla 2x2.}
}
\value{
Numero total de observaciones.
}
\description{
Calcula el numero total de observaciones de una tabla 2x2.
}
