% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metricas_derivadas.R
\name{indice_balanced_accuracy}
\alias{indice_balanced_accuracy}
\title{Calcular la balanced accuracy}
\usage{
indice_balanced_accuracy(
  tab,
  se_method = c("analytic", "bootstrap", "none"),
  B = 1000,
  seed = 123
)
}
\arguments{
\item{tab}{Matriz o tabla 2x2 con frecuencias observadas.}

\item{se_method}{Método para calcular el error estándar: "analytic", "bootstrap" o "none".}

\item{B}{Número de remuestreos bootstrap.}

\item{seed}{Semilla para reproducibilidad.}
}
\value{
Lista con statistic y se.
}
\description{
Calcula la balanced accuracy, definida como la media entre la sensibilidad
y la especificidad.
}
\examples{
tab <- matrix(c(50, 10,
                5, 35),
              nrow = 2,
              byrow = TRUE)
indice_balanced_accuracy(tab)

}
