% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metricas_derivadas.R
\name{indice_f1}
\alias{indice_f1}
\title{Calcular el F1-score}
\usage{
indice_f1(tab, se_method = c("bootstrap", "none"), B = 1000, seed = 123)
}
\arguments{
\item{tab}{Matriz o tabla 2x2 con frecuencias observadas.}

\item{se_method}{Método para calcular el error estándar: "bootstrap" o "none".}

\item{B}{Número de remuestreos bootstrap.}

\item{seed}{Semilla para reproducibilidad.}
}
\value{
Lista con statistic y se.
}
\description{
Calcula el F1-score, definido como la media armónica entre precision
y recall. En términos de la tabla 2x2 puede escribirse como
2TP/(2TP+FP+FN).
}
\examples{
tab <- matrix(c(50, 10,
                5, 35),
              nrow = 2,
              byrow = TRUE)
indice_f1(tab, B = 100)

}
