% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metricas_basicas.R
\name{indice_fnr}
\alias{indice_fnr}
\title{Calcular la tasa de falsos negativos}
\usage{
indice_fnr(
  tab,
  se_method = c("analytic", "bootstrap", "none"),
  B = 1000,
  seed = 123
)
}
\arguments{
\item{tab}{Matriz o tabla 2x2 con frecuencias observadas.}

\item{se_method}{Método para calcular el error estándar: "analytic", "bootstrap" o "none".}

\item{B}{Número de remuestreos bootstrap.}

\item{seed}{Semilla para reproducibilidad.}
}
\value{
Lista con statistic y se.
}
\description{
Calcula la false negative rate (FNR), definida como FN/(FN+TP).
}
\examples{
tab <- matrix(c(50, 10,
                5, 35),
              nrow = 2,
              byrow = TRUE)
indice_fnr(tab)

}
