% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metricas_clinicas_y_asociacion.R
\name{indice_lr_neg}
\alias{indice_lr_neg}
\title{Calcular el likelihood ratio negativo}
\usage{
indice_lr_neg(tab, se_method = c("bootstrap", "none"), B = 1000, seed = 123)
}
\arguments{
\item{tab}{Matriz o tabla 2x2 con frecuencias observadas.}

\item{se_method}{Método para calcular el error estándar: "bootstrap" o "none".}

\item{B}{Número de remuestreos bootstrap.}

\item{seed}{Semilla para reproducibilidad.}
}
\value{
Lista con statistic y se.
}
\description{
Calcula el likelihood ratio negativo (LR-), definido como
(1 - sensibilidad) / especificidad.
}
\examples{
tab <- matrix(c(50, 10,
                5, 35),
              nrow = 2,
              byrow = TRUE)
indice_lr_neg(tab, B = 100)

}
