% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metricas_derivadas.R
\name{indice_markedness}
\alias{indice_markedness}
\title{Calcular la markedness}
\usage{
indice_markedness(
  tab,
  se_method = c("analytic", "bootstrap", "none"),
  B = 1000,
  seed = 123
)
}
\arguments{
\item{tab}{Matriz o tabla 2x2 con frecuencias observadas.}

\item{se_method}{Método para calcular el error estándar: "analytic", "bootstrap" o "none".}

\item{B}{Número de remuestreos bootstrap.}

\item{seed}{Semilla para reproducibilidad.}
}
\value{
Lista con statistic y se.
}
\description{
Calcula la markedness, definida como PPV + NPV - 1.
Es una medida dual del índice de Youden.
}
\examples{
tab <- matrix(c(50, 10,
                5, 35),
              nrow = 2,
              byrow = TRUE)
indice_markedness(tab)

}
