% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metricas_acuerdo_corregido_azar.R
\name{indice_scott_pi}
\alias{indice_scott_pi}
\title{Calcular Scott Pi}
\usage{
indice_scott_pi(tab, se_method = c("bootstrap", "none"), B = 1000, seed = 123)
}
\arguments{
\item{tab}{Matriz o tabla 2x2 con frecuencias observadas.}

\item{se_method}{Método para calcular el error estándar: "bootstrap" o "none".}

\item{B}{Número de remuestreos bootstrap.}

\item{seed}{Semilla para reproducibilidad.}
}
\value{
Lista con statistic y se.
}
\description{
Calcula Scott Pi, una medida de acuerdo corregida por azar basada en una
estimación común de la prevalencia marginal.
}
\examples{
tab <- matrix(c(50, 10,
                5, 35),
              nrow = 2,
              byrow = TRUE)
indice_scott_pi(tab, B = 100)

}
