% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metricas_basicas.R
\name{indice_sensibilidad}
\alias{indice_sensibilidad}
\title{Calcular la sensibilidad}
\usage{
indice_sensibilidad(
  tab,
  se_method = c("analytic", "bootstrap", "none"),
  B = 1000,
  seed = 123
)
}
\arguments{
\item{tab}{Matriz o tabla 2x2 con la convención TP, FN, FP, TN.}

\item{se_method}{Método para calcular el error estándar: "analytic", "bootstrap" o "none".}

\item{B}{Número de remuestreos bootstrap.}

\item{seed}{Semilla para reproducibilidad.}
}
\value{
Una lista con dos elementos: statistic y se.
}
\description{
Calcula la sensibilidad, también conocida como recall o true positive rate.
}
\examples{
tab <- matrix(c(50, 10,
                5, 35),
              nrow = 2,
              byrow = TRUE)

indice_sensibilidad(tab)

}
