% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metricas_derivadas.R
\name{indice_youden}
\alias{indice_youden}
\title{Calcular el índice de Youden}
\usage{
indice_youden(
  tab,
  se_method = c("analytic", "bootstrap", "none"),
  B = 1000,
  seed = 123
)
}
\arguments{
\item{tab}{Matriz o tabla 2x2 con frecuencias observadas.}

\item{se_method}{Método para calcular el error estándar: "analytic", "bootstrap" o "none".}

\item{B}{Número de remuestreos bootstrap.}

\item{seed}{Semilla para reproducibilidad.}
}
\value{
Lista con statistic y se.
}
\description{
Calcula el índice de Youden, definido como sensibilidad + especificidad - 1.
También se conoce como informedness o bookmaker informedness.
}
\examples{
tab <- matrix(c(50, 10,
                5, 35),
              nrow = 2,
              byrow = TRUE)
indice_youden(tab)

}
