% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metricas_clinicas_y_asociacion.R
\name{indice_yule_q}
\alias{indice_yule_q}
\title{Calcular Yule Q}
\usage{
indice_yule_q(tab, se_method = c("bootstrap", "none"), B = 1000, seed = 123)
}
\arguments{
\item{tab}{Matriz o tabla 2x2 con frecuencias observadas.}

\item{se_method}{Método para calcular el error estándar: "bootstrap" o "none".}

\item{B}{Número de remuestreos bootstrap.}

\item{seed}{Semilla para reproducibilidad.}
}
\value{
Lista con statistic y se.
}
\description{
Calcula la medida de asociación Yule Q, definida como
(TP * TN - FP * FN) / (TP * TN + FP * FN).
}
\examples{
tab <- matrix(c(50, 10,
                5, 35),
              nrow = 2,
              byrow = TRUE)
indice_yule_q(tab, B = 100)

}
