% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metricas_score_based.R
\name{pr_points}
\alias{pr_points}
\title{Obtener puntos de la curva Precision-Recall}
\usage{
pr_points(y_true, y_score)
}
\arguments{
\item{y_true}{Vector con valores reales codificados como 0 y 1.}

\item{y_score}{Vector con scores o probabilidades predichas.}
}
\value{
Data frame con umbral, recall y precision.
}
\description{
Calcula los puntos de la curva precision-recall al variar el umbral
de clasificación.
}
\examples{
y_true <- c(1, 1, 0, 0)
y_score <- c(0.9, 0.7, 0.4, 0.2)
pr_points(y_true, y_score)

}
