% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliares.R
\name{safe_divide}
\alias{safe_divide}
\title{Division segura}
\usage{
safe_divide(num, den)
}
\arguments{
\item{num}{Numerador.}

\item{den}{Denominador.}
}
\value{
Resultado de la division o NA si no esta definida.
}
\description{
Realiza una division controlando denominadores nulos o valores perdidos.
}
